requirements.txt
streamlit
langchain
langchain_chroma
langchain_community
langchain_core
langchain_groq
langchain_huggingface
langchain_text_splitters
python-dotenv
chromadb